requirements.txt
flask
opencv-python
numpy
python-telegram-bot
